Post ready message to console

+++ console/consoleinputhandler.cpp
@@ -21,40 +21,43 @@ static std::pair<intptr_t,int> line_length(const uint32_t* begin,const uint32_t*
 void ConsoleInputHandler::key(int scancode,GLFWmm::WindowBase::Action action
 	,unsigned int modifiers)
 	{
-	switch(scancode)
+	if(m_ready)
 		{
-		case 22:
-			if(m_input_buffer.length()!=0 && action!=GLFWmm::WindowBase::Action::RELEASE)
-				{
-				auto ch=*(m_input_buffer.end() - 1);
-				m_input_buffer.truncate();
-				if(ch=='\n')
+		switch(scancode)
+			{
+			case 22:
+				if(m_input_buffer.length()!=0 && action!=GLFWmm::WindowBase::Action::RELEASE)
 					{
-					auto l=line_length(m_input_buffer.begin(),m_input_buffer.end());
-					r_con->eraseLinefeed(l.first + l.second);
+					auto ch=*(m_input_buffer.end() - 1);
+					m_input_buffer.truncate();
+					if(ch=='\n')
+						{
+						auto l=line_length(m_input_buffer.begin(),m_input_buffer.end());
+						r_con->eraseLinefeed(l.first + l.second);
+						}
+					else
+						{r_con->erase();}
 					}
-				else
-					{r_con->erase();}
-				}
-			break;
+				break;
 
-		case 36:
-			if(action!=GLFWmm::WindowBase::Action::RELEASE)
-				{
-				r_con->write('\n');
-				if(modifiers==0)
+			case 36:
+				if(action!=GLFWmm::WindowBase::Action::RELEASE)
 					{
-					fprintf(stderr,"Invoke command\n");
-					m_input_buffer.clear();
-					r_con->writeUTF8("»");
+					r_con->write('\n');
+					if(modifiers==0)
+						{
+						fprintf(stderr,"Invoke command\n");
+						m_input_buffer.clear();
+						r_con->writeUTF8("»");
+						}
+					else
+						{m_input_buffer.append('\n');}
 					}
-				else
-					{m_input_buffer.append('\n');}
-				}
-			break;
+				break;
 
-		default:
-			break;
-		//	fprintf(stderr,"%d\n",scancode);
+			default:
+				break;
+			//	fprintf(stderr,"%d\n",scancode);
+			}
 		}
 	}
